+++ SendBird.d.ts
@@ -1,7 +1,7 @@
 /**
- * Type Definitions for SendBird SDK v3.0.131
+ * Type Definitions for Sendbird SDK v3.0.131
  * homepage: https://sendbird.com/
- * git: https://github.com/sendbird/SendBird-SDK-JavaScript
+ * git: https://github.com/sendbird/Sendbird-SDK-JavaScript
  */
 
 export = SendBird;
@@ -48,7 +48,7 @@ declare namespace SendBird {
 
   type pushTokens = {
     deviceTokens: Array<string>;
-    type: 'gcm' | 'apns' | 'apns_voip';
+    type: "gcm" | "apns" | "apns_voip";
     hasMore: boolean;
     token: string;
   };
@@ -142,12 +142,12 @@ declare namespace SendBird {
     unregisterAPNSPushTokenAllForCurrentUser(callback?: commonCallback): void;
 
     unregisterPushTokenAllForCurrentUser(callback?: commonCallback): void; // This removes all push tokens including APNS/GCM
-    getMyPushTokensByToken(token: string, type: 'gcm' | 'apns' | 'apns_voip', callback: getMyPushTokensHandler): void;
+    getMyPushTokensByToken(token: string, type: "gcm" | "apns" | "apns_voip", callback: getMyPushTokensHandler): void;
 
     setPushTemplate(templateName: string, callback?: pushSettingCallback): void;
     getPushTemplate(callback?: pushSettingCallback): void;
 
-    setPushTriggerOption(pushTriggerOption: 'all' | 'mention_only' | 'off', callback: commonCallback): void;
+    setPushTriggerOption(pushTriggerOption: "all" | "mention_only" | "off", callback: commonCallback): void;
     getPushTriggerOption(callback: getPushTriggerOptionCallback): void;
 
     setDoNotDisturb(
@@ -193,10 +193,7 @@ declare namespace SendBird {
     getGroupChannelCount(memberStateFilter: string, callback: commonCallback): void;
 
     getUnreadItemCount(keys: Array<string>, callback: commonCallback): void;
-    getTotalUnreadMessageCount(
-      groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams,
-      callback: groupChannelCountCallback
-    ): void;
+    getTotalUnreadMessageCount(groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams, callback: groupChannelCountCallback): void;
     getTotalUnreadMessageCount(callback: groupChannelCountCallback): void;
     getTotalUnreadMessageCount(channelCustomTypes: Array<string>, callback: groupChannelCountCallback): void;
     getTotalUnreadChannelCount(callback: groupChannelCountCallback): void;
@@ -206,41 +203,15 @@ declare namespace SendBird {
     getSubscribedCustomTypeUnreadMessageCount(customType: string): number;
 
     // DEPRECATED
-    getMyGroupChannelChangeLogsByToken(
-      token: string,
-      customTypes: Array<string>,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
+    getMyGroupChannelChangeLogsByToken(token: string, customTypes: Array<string>, callback: getGroupChannelChangeLogsHandler): void;
     // DEPRECATED
-    getMyGroupChannelChangeLogsByToken(
-      token: string,
-      customTypes: Array<string>,
-      includeEmpty: boolean,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
-    getMyGroupChannelChangeLogsByToken(
-      token: string,
-      params: GroupChannelChangeLogsParams,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
+    getMyGroupChannelChangeLogsByToken(token: string, customTypes: Array<string>, includeEmpty: boolean, callback: getGroupChannelChangeLogsHandler): void;
+    getMyGroupChannelChangeLogsByToken(token: string, params: GroupChannelChangeLogsParams, callback: getGroupChannelChangeLogsHandler): void;
     // DEPRECATED
-    getMyGroupChannelChangeLogsByTimestamp(
-      ts: number,
-      customTypes: Array<string>,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
+    getMyGroupChannelChangeLogsByTimestamp(ts: number, customTypes: Array<string>, callback: getGroupChannelChangeLogsHandler): void;
     // DEPRECATED
-    getMyGroupChannelChangeLogsByTimestamp(
-      ts: number,
-      customTypes: Array<string>,
-      includeEmpty: boolean,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
-    getMyGroupChannelChangeLogsByTimestamp(
-      ts: number,
-      params: GroupChannelChangeLogsParams,
-      callback: getGroupChannelChangeLogsHandler
-    ): void;
+    getMyGroupChannelChangeLogsByTimestamp(ts: number, customTypes: Array<string>, includeEmpty: boolean, callback: getGroupChannelChangeLogsHandler): void;
+    getMyGroupChannelChangeLogsByTimestamp(ts: number, params: GroupChannelChangeLogsParams, callback: getGroupChannelChangeLogsHandler): void;
 
     getAllEmoji(callback: getEmojiContainerHandler): void;
     getEmojiCategory(categoryId: number, callback: getEmojiCategoryHandler): void;
@@ -281,7 +252,7 @@ declare namespace SendBird {
     channelCustomType: string;
     messageTimestampFrom: number;
     messageTimestampTo: number;
-    order: 'score' | 'ts';
+    order: "score" | "ts";
   }
 
   interface UserEventHandlerStatic {
@@ -344,7 +315,7 @@ declare namespace SendBird {
     channelUrl: string;
     channelType: string;
     messageId: number;
-    messageType: 'base' | 'user' | 'file' | 'admin';
+    messageType: "base" | "user" | "file" | "admin";
     data: string;
     customType: string;
     metaArrays: Array<MessageMetaArray>;
@@ -369,11 +340,7 @@ declare namespace SendBird {
     serialize(): Object;
     getMetaArraysByKeys(keys: Array<string>): Array<MessageMetaArray>;
     applyReactionEvent(event: ReactionEvent): void;
-    getThreadedMessagesByTimestamp(
-      timestamp: number,
-      params: ThreadedMessageListParams,
-      callback: messageListCallback
-    ): void;
+    getThreadedMessagesByTimestamp(timestamp: number, params: ThreadedMessageListParams, callback: messageListCallback): void;
     applyThreadInfoUpdateEvent(threadInfoUpdateEvent: ThreadInfoUpdateEvent): boolean;
 
     metaArray: Object; // DEPRECATED
@@ -381,7 +348,7 @@ declare namespace SendBird {
   }
 
   interface AdminMessage extends BaseMessageInstance {
-    messageType: 'admin';
+    messageType: "admin";
     message: string;
     translations: Object;
   }
@@ -393,7 +360,7 @@ declare namespace SendBird {
   interface GroupChannelTotalUnreadMessageCountParams {
     new (): GroupChannelTotalUnreadMessageCountParams;
     channelCustomTypesFilter: Array<string>;
-    superChannelFilter: 'all' | 'super' | 'nonsuper';
+    superChannelFilter: "all" | "super" | "nonsuper";
   }
 
   interface Emoji {
@@ -420,22 +387,22 @@ declare namespace SendBird {
     customType: string;
     targetLanguages: Array<string>; // DEPRECATED
     translationTargetLanguages: Array<string>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUserIds: Array<string>;
     mentionedUsers: Array<User>;
     metaArrays: Array<MessageMetaArray>;
     metaArrayKeys: Array<string>; // DEPRECATED
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
     parentMessageId: number;
   }
   interface UserMessage extends BaseMessageInstance {
-    messageType: 'user';
+    messageType: "user";
     message: string;
     sender: Sender;
     reqId: string;
     translations: Object;
-    requestState: 'none' | 'pending' | 'failed' | 'succeeded'; // DEPRECATED
-    sendingStatus: 'none' | 'pending' | 'failed' | 'canceled' | 'succeeded';
+    requestState: "none" | "pending" | "failed" | "succeeded"; // DEPRECATED
+    sendingStatus: "none" | "pending" | "failed" | "canceled" | "succeeded";
     requestedMentionUserIds: Array<string>;
     errorCode: number;
     messageSurvivalSeconds: number;
@@ -456,16 +423,16 @@ declare namespace SendBird {
     data: string;
     customType: string;
     thumbnailSizes: Array<ThumbnailSize>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUserIds: Array<string>;
     mentionedUsers: Array<User>;
     metaArrays: Array<MessageMetaArray>;
     metaArrayKeys: Array<string>; // DEPRECATED
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
     parentMessageId: number;
   }
   interface FileMessage extends BaseMessageInstance {
-    messageType: 'file';
+    messageType: "file";
     sender: Sender;
     reqId: string;
     plainUrl: string;
@@ -474,8 +441,8 @@ declare namespace SendBird {
     size: number;
     type: string;
     thumbnails: Array<ThumbnailObject>;
-    requestState: 'none' | 'pending' | 'failed' | 'succeeded'; // DEPRECATED
-    sendingStatus: 'none' | 'pending' | 'failed' | 'canceled' | 'succeeded';
+    requestState: "none" | "pending" | "failed" | "succeeded"; // DEPRECATED
+    sendingStatus: "none" | "pending" | "failed" | "canceled" | "succeeded";
     requestedMentionUserIds: Array<string>;
     errorCode: number;
     messageSurvivalSeconds: number;
@@ -583,8 +550,8 @@ declare namespace SendBird {
   }
 
   interface Member extends User {
-    state: 'invited' | 'joined';
-    role: 'none' | 'operator';
+    state: "invited" | "joined";
+    role: "none" | "operator";
     isMuted: boolean;
     isBlockedByMe: boolean;
     isBlockingMe: boolean;
@@ -657,16 +624,8 @@ declare namespace SendBird {
       callback: getMessageChangeLogsHandler
     ): void;
 
-    getMessageChangeLogsSinceToken(
-      token: string,
-      params: MessageChangeLogsParams,
-      callback: getMessageChangeLogsHandler
-    ): void;
-    getMessageChangeLogsSinceTimestamp(
-      timestamp: number,
-      params: MessageChangeLogsParams,
-      callback: getMessageChangeLogsHandler
-    ): void;
+    getMessageChangeLogsSinceToken(token: string, params: MessageChangeLogsParams, callback: getMessageChangeLogsHandler): void;
+    getMessageChangeLogsSinceTimestamp(timestamp: number, params: MessageChangeLogsParams, callback: getMessageChangeLogsHandler): void;
 
     getMyMutedInfo(callback: getMyMutedInfoHandler): void;
     createOperatorListQuery(): OperatorListQuery;
@@ -943,30 +902,9 @@ declare namespace SendBird {
     sendFileMessage(file: File, callback: messageCallback): FileMessage;
     sendFileMessage(file: File, data: string, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(file: File, data: string, customType: string, callback: messageCallback): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      data: string,
-      customType: string,
-      thumbnailSizes: Array<ThumbnailSize>,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      customType: string,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, data: string, customType: string, thumbnailSizes: Array<ThumbnailSize>, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, name: string, type: string, size: number, data: string, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, name: string, type: string, size: number, data: string, customType: string, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(
       file: File,
       name: string,
@@ -978,25 +916,10 @@ declare namespace SendBird {
       callback: messageCallback
     ): FileMessage;
 
-    sendFileMessage(
-      fileMessageParams: FileMessageParams,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage;
+    sendFileMessage(fileMessageParams: FileMessageParams, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage;
     sendFileMessage(file: File, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage;
-    sendFileMessage(
-      file: File,
-      data: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: File,
-      data: string,
-      customType: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, data: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: File, data: string, customType: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(
       file: File,
       data: string,
@@ -1039,38 +962,12 @@ declare namespace SendBird {
     sendFileMessage(file: string, callback: messageCallback): FileMessage;
     sendFileMessage(file: string, data: string, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(file: string, data: string, customType: string, callback: messageCallback): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      name: string,
-      type: string,
-      size: number,
-      data: string,
-      customType: string,
-      callback: messageCallback
-    ): FileMessage;
+    sendFileMessage(file: string, name: string, type: string, size: number, data: string, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: string, name: string, type: string, size: number, data: string, customType: string, callback: messageCallback): FileMessage;
 
     sendFileMessage(file: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      data: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
-    sendFileMessage(
-      file: string,
-      data: string,
-      customType: string,
-      progressHandler: fileUploadprogressHandler,
-      callback: messageCallback
-    ): FileMessage; // DEPRECATED
+    sendFileMessage(file: string, data: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
+    sendFileMessage(file: string, data: string, customType: string, progressHandler: fileUploadprogressHandler, callback: messageCallback): FileMessage; // DEPRECATED
     sendFileMessage(
       file: string,
       name: string,
@@ -1091,44 +988,25 @@ declare namespace SendBird {
       callback: messageCallback
     ): FileMessage; // DEPRECATED
 
-    sendFileMessages(
-      fileMessageParamsList: Array<FileMessageParams>,
-      callbackObject: fileMessagesCallbackObject
-    ): Array<FileMessage>;
+    sendFileMessages(fileMessageParamsList: Array<FileMessageParams>, callbackObject: fileMessagesCallbackObject): Array<FileMessage>;
 
     /** UserMessage  */
     sendUserMessage(userMessageParams: UserMessageParams, callback: messageCallback): UserMessage;
     sendUserMessage(message: string, callback: messageCallback): UserMessage;
     sendUserMessage(message: string, data: string, callback: messageCallback): UserMessage; // DEPRECATED
     sendUserMessage(message: string, data: string, customType: string, callback: messageCallback): UserMessage; // DEPRECATED
-    sendUserMessage(
-      message: string,
-      data: string,
-      customType: string,
-      translationTargetLanguages: Array<string>,
-      callback: messageCallback
-    ): UserMessage;
+    sendUserMessage(message: string, data: string, customType: string, translationTargetLanguages: Array<string>, callback: messageCallback): UserMessage;
 
     resendUserMessage(userMessage: UserMessage, callback: messageCallback): void;
     resendFileMessage(fileMessage: FileMessage, callback: messageCallback): void;
     resendFileMessage(fileMessage: FileMessage, file: Blob, callback: messageCallback): void;
 
-    translateUserMessage(
-      message: UserMessage,
-      translationTargetLanguages: Array<string>,
-      callback: messageCallback
-    ): void;
+    translateUserMessage(message: UserMessage, translationTargetLanguages: Array<string>, callback: messageCallback): void;
 
     /** Edit Message  */
     updateFileMessage(messageId: number, data: string, customType: string, callback: messageCallback): void;
     updateFileMessage(messageId: number, fileMessageParams: FileMessageParams, callback: messageCallback): void;
-    updateUserMessage(
-      messageId: number,
-      message: string,
-      data: string,
-      customType: string,
-      callback: messageCallback
-    ): void;
+    updateUserMessage(messageId: number, message: string, data: string, customType: string, callback: messageCallback): void;
     updateUserMessage(messageId: number, userMessageParams: UserMessageParams, callback: messageCallback): void;
     deleteMessage(message: FileMessage | UserMessage, callback: commonCallback): void;
     cancelUploadingFileMessage(messageReqId: string, callback: cancelUploadingFileMessageCallback): boolean;
@@ -1139,11 +1017,7 @@ declare namespace SendBird {
 
     /** Reaction */
     addReaction(message: UserMessage | FileMessage | AdminMessage, key: string, callback: reactionEventCallback): void;
-    deleteReaction(
-      message: UserMessage | FileMessage | AdminMessage,
-      key: string,
-      callback: reactionEventCallback
-    ): void;
+    deleteReaction(message: UserMessage | FileMessage | AdminMessage, key: string, callback: reactionEventCallback): void;
 
     /** MetaData */
     createMetaData(metaDataMap: Object, callback: commonCallback): void;
@@ -1166,64 +1040,26 @@ declare namespace SendBird {
     deleteAllMetaCounters(callback: commonCallback): void;
 
     /** MessageMetaArray */
-    createMessageMetaArrayKeys(
-      message: UserMessage | FileMessage | AdminMessage,
-      keys: Array<string>,
-      callback: commonCallback
-    ): void;
-    deleteMessageMetaArrayKeys(
-      message: UserMessage | FileMessage | AdminMessage,
-      keys: Array<string>,
-      callback: commonCallback
-    ): void;
-    addMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Array<MessageMetaArray>,
-      callback: commonCallback
-    ): void;
-    removeMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Array<MessageMetaArray>,
-      callback: commonCallback
-    ): void;
+    createMessageMetaArrayKeys(message: UserMessage | FileMessage | AdminMessage, keys: Array<string>, callback: commonCallback): void;
+    deleteMessageMetaArrayKeys(message: UserMessage | FileMessage | AdminMessage, keys: Array<string>, callback: commonCallback): void;
+    addMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Array<MessageMetaArray>, callback: commonCallback): void;
+    removeMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Array<MessageMetaArray>, callback: commonCallback): void;
 
-    addMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Object,
-      callback: commonCallback
-    ): void; // DEPRECATED
-    removeMessageMetaArrayValues(
-      message: UserMessage | FileMessage | AdminMessage,
-      data: Object,
-      callback: commonCallback
-    ): void; // DEPRECATED
+    addMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Object, callback: commonCallback): void; // DEPRECATED
+    removeMessageMetaArrayValues(message: UserMessage | FileMessage | AdminMessage, data: Object, callback: commonCallback): void; // DEPRECATED
 
     /** Report */
     report(category: string, description: string, callback: commonCallback): void;
     reportUser(user: User, category: string, description: string, callback: commonCallback): void;
-    reportMessage(
-      message: UserMessage | FileMessage | AdminMessage,
-      category: string,
-      description: string,
-      callback: commonCallback
-    ): void;
+    reportMessage(message: UserMessage | FileMessage | AdminMessage, category: string, description: string, callback: commonCallback): void;
   }
 
-  type messageListCallback = (
-    messageList: Array<UserMessage | FileMessage | AdminMessage>,
-    error: SendBirdError
-  ) => void;
+  type messageListCallback = (messageList: Array<UserMessage | FileMessage | AdminMessage>, error: SendBirdError) => void;
   interface MessageListQuery {
     // DEPRECATED
     next(messageTimestamp: number, limit: number, reverse: boolean, callback: messageListCallback): void;
     prev(messageTimestamp: number, limit: number, reverse: boolean, callback: messageListCallback): void;
-    load(
-      messageTimestamp: number,
-      prevLimit: number,
-      nextLimit: number,
-      reverse: boolean,
-      callback: messageListCallback
-    ): void;
+    load(messageTimestamp: number, prevLimit: number, nextLimit: number, reverse: boolean, callback: messageListCallback): void;
   }
 
   type threadedMessageList = {
@@ -1271,13 +1107,7 @@ declare namespace SendBird {
     createBannedUserListQuery(): BannedUserListQuery;
 
     updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: openChannelCallback): void;
-    updateChannel(
-      name: string,
-      coverUrlOrImageFile: string | File,
-      data: string,
-      operatorUserIds: Array<string> | string,
-      callback: openChannelCallback
-    ): void; // DEPRECATED
+    updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, operatorUserIds: Array<string> | string, callback: openChannelCallback): void; // DEPRECATED
     updateChannel(
       name: string,
       coverUrlOrImageFile: string | File,
@@ -1392,13 +1222,7 @@ declare namespace SendBird {
 
     createChannel(callback: openChannelCallback): void;
     createChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: openChannelCallback): void; // DEPRECATED
-    createChannel(
-      name: string,
-      coverUrlOrImageFile: string | File,
-      data: string,
-      operatorUserIds: Array<string> | string,
-      callback: openChannelCallback
-    ): void; // DEPRECATED
+    createChannel(name: string, coverUrlOrImageFile: string | File, data: string, operatorUserIds: Array<string> | string, callback: openChannelCallback): void; // DEPRECATED
     createChannel(
       name: string,
       coverUrlOrImageFile: string | File,
@@ -1446,7 +1270,7 @@ declare namespace SendBird {
     scheduledId: number;
     scheduledDateTimeString: string;
     scheduledTimezone: string;
-    status: 'scheduled' | 'sent' | 'canceled' | 'failed';
+    status: "scheduled" | "sent" | "canceled" | "failed";
     createdAt: number;
     updatedAt: number;
     channelUrl: string;
@@ -1456,9 +1280,9 @@ declare namespace SendBird {
     data: string;
     metaArray: Object; // DEPRECATED
     metaArrays: Array<MessageMetaArray>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUsers: Array<User>;
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
     translationTargetLanguages: Array<string>;
     errorMessage: string;
     errorCode: number;
@@ -1505,10 +1329,10 @@ declare namespace SendBird {
     customType: string;
     translationTargetLanguages: Array<string>;
     metaArrayKeys: Array<string>;
-    mentionType: 'users' | 'channel';
+    mentionType: "users" | "channel";
     mentionedUserIds: Array<string>;
     mentionedUsers: Array<User>;
-    pushNotificationDeliveryOption: 'default' | 'suppress';
+    pushNotificationDeliveryOption: "default" | "suppress";
     year: number;
     month: number;
     day: number;
@@ -1548,7 +1372,7 @@ declare namespace SendBird {
     messageId: string;
     userId: string;
     key: string;
-    operation: 'add' | 'delete';
+    operation: "add" | "delete";
     updatedAt: number;
   }
 
@@ -1587,10 +1411,7 @@ declare namespace SendBird {
   type groupChannelCallback = (groupChannel: GroupChannel, error: SendBirdError) => void;
   type distinctGroupChannelCallback = (response: DistinctGroupChannelResponse, error: SendBirdError) => void;
   type getPushPreferenceCallback = (isPushOn: boolean, error: SendBirdError) => void;
-  type getPushTriggerOptionCallback = (
-    pushTriggerOption: 'all' | 'mention_only' | 'off' | 'default',
-    error: SendBirdError
-  ) => void;
+  type getPushTriggerOptionCallback = (pushTriggerOption: "all" | "mention_only" | "off" | "default", error: SendBirdError) => void;
   interface GroupChannel extends BaseChannel {
     isHidden: boolean;
     isDistinct: boolean;
@@ -1598,7 +1419,7 @@ declare namespace SendBird {
     isBroadcast: boolean;
     isPublic: boolean;
     isPushEnabled: boolean; // DEPRECATED
-    myPushTriggerOption: 'default' | 'all' | 'mention_only' | 'off';
+    myPushTriggerOption: "default" | "all" | "mention_only" | "off";
     myCountPreference: string;
     lastMessage: UserMessage | FileMessage | AdminMessage;
     unreadMessageCount: number;
@@ -1606,13 +1427,13 @@ declare namespace SendBird {
     members: Array<Member>;
     memberCount: number;
     joinedMemberCount: number;
-    myMemberState: 'none' | 'joined' | 'invited';
-    myRole: 'operator' | 'none';
-    myMutedState: 'muted' | 'unmuted';
+    myMemberState: "none" | "joined" | "invited";
+    myRole: "operator" | "none";
+    myMutedState: "muted" | "unmuted";
     inviter: User;
     invitedAt: number;
     isAccessCodeRequired: boolean;
-    hiddenState: 'unhidden' | 'hidden_allow_auto_unhide' | 'hidden_prevent_auto_unhide';
+    hiddenState: "unhidden" | "hidden_allow_auto_unhide" | "hidden_prevent_auto_unhide";
     isDiscoverable: boolean;
     myLastRead: number;
     messageOffsetTimestamp: number;
@@ -1625,13 +1446,7 @@ declare namespace SendBird {
 
     updateChannel(groupChannelParams: GroupChannelParams, callback: groupChannelCallback): void;
     updateChannel(name: string, coverUrlOrImageFile: string | File, data: string, callback: groupChannelCallback): void;
-    updateChannel(
-      isDistinct: boolean,
-      name: string,
-      coverUrlOrImageFile: string | File,
-      data: string,
-      callback: groupChannelCallback
-    ): void; // DEPRECATED
+    updateChannel(isDistinct: boolean, name: string, coverUrlOrImageFile: string | File, data: string, callback: groupChannelCallback): void; // DEPRECATED
     updateChannel(
       isDistinct: boolean,
       name: string,
@@ -1676,15 +1491,9 @@ declare namespace SendBird {
 
     setPushPreference(pushOn: boolean, callback: commonCallback): void; // DEPRECATED
     getPushPreference(callback: getPushPreferenceCallback): void; // DEPRECATED
-    setMyPushTriggerOption(
-      pushTriggerOption: 'all' | 'mention_only' | 'off' | 'default',
-      callback: getPushTriggerOptionCallback
-    ): void;
+    setMyPushTriggerOption(pushTriggerOption: "all" | "mention_only" | "off" | "default", callback: getPushTriggerOptionCallback): void;
     getMyPushTriggerOption(callback: getPushTriggerOptionCallback): void;
-    setMyCountPreference(
-      preference: 'all' | 'unread_message_count_only' | 'unread_mention_count_only' | 'off',
-      callback: commonCallback
-    ): void;
+    setMyCountPreference(preference: "all" | "unread_message_count_only" | "unread_mention_count_only" | "off", callback: commonCallback): void;
 
     createMemberListQuery(): GroupChannelMemberListQuery;
     createBannedUserListQuery(): BannedUserListQuery;
@@ -1704,10 +1513,7 @@ declare namespace SendBird {
     freeze(callback: commonCallback): void;
     unfreeze(callback: commonCallback): void;
 
-    registerScheduledUserMessage(
-      scheduledUserMessageParams: ScheduledUserMessageParams,
-      callback: scheduledUserMessageCallback
-    ): void;
+    registerScheduledUserMessage(scheduledUserMessageParams: ScheduledUserMessageParams, callback: scheduledUserMessageCallback): void;
   }
 
   type groupChannelCountCallback = (count: number, error: SendBirdError) => void;
@@ -1718,10 +1524,7 @@ declare namespace SendBird {
     createPublicGroupChannelListQuery(): PublicGroupChannelListQuery;
 
     getUnreadItemCount(keys: Array<string>, callback: commonCallback): void; // DEPRECATED
-    getTotalUnreadMessageCount(
-      groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams,
-      callback: groupChannelCountCallback
-    ): void; // DEPRECATED
+    getTotalUnreadMessageCount(groupChannelTotalUnreadMessageCountParams: GroupChannelTotalUnreadMessageCountParams, callback: groupChannelCountCallback): void; // DEPRECATED
     getTotalUnreadMessageCount(callback: groupChannelCountCallback): void; // DEPRECATED
     getTotalUnreadMessageCount(channelCustomTypes: Array<string>, callback: groupChannelCountCallback): void; // DEPRECATED
     getTotalUnreadChannelCount(callback: groupChannelCountCallback): void; // DEPRECATED
@@ -1747,19 +1550,11 @@ declare namespace SendBird {
       customType: string,
       callback: groupChannelCallback
     ): void;
-    createDistinctChannelIfNotExist(
-      groupChannelParams: GroupChannelParams,
-      callback: distinctGroupChannelCallback
-    ): void;
+    createDistinctChannelIfNotExist(groupChannelParams: GroupChannelParams, callback: distinctGroupChannelCallback): void;
 
     createChannelWithUserIds(userIds: Array<string>, callback: groupChannelCallback): void;
     createChannelWithUserIds(userIds: Array<string>, isDistinct: boolean, callback: groupChannelCallback): void; // DEPRECATED
-    createChannelWithUserIds(
-      userIds: Array<string>,
-      isDistinct: boolean,
-      customType: string,
-      callback: groupChannelCallback
-    ): void; // DEPRECATED
+    createChannelWithUserIds(userIds: Array<string>, isDistinct: boolean, customType: string, callback: groupChannelCallback): void; // DEPRECATED
     createChannelWithUserIds(
       userIds: Array<string>,
       isDistinct: boolean,
@@ -1789,10 +1584,10 @@ declare namespace SendBird {
     limit: number;
     hasNext: boolean;
     isLoading: boolean;
-    order: 'member_nickname_alphabetical' | 'operator_then_member_alphabetical';
-    mutedMemberFilter: 'all' | 'muted' | 'unmuted';
-    operatorFilter: 'all' | 'operator' | 'nonoperator'; // DEPRECATED
-    memberStateFilter: 'all' | 'joined_only' | 'invited_only' | 'invited_by_friend' | 'invited_by_non_friend';
+    order: "member_nickname_alphabetical" | "operator_then_member_alphabetical";
+    mutedMemberFilter: "all" | "muted" | "unmuted";
+    operatorFilter: "all" | "operator" | "nonoperator"; // DEPRECATED
+    memberStateFilter: "all" | "joined_only" | "invited_only" | "invited_by_friend" | "invited_by_non_friend";
     nicknameStartsWithFilter: string;
 
     next(callback: groupChannelMemberListQueryCallback): void;
@@ -1807,25 +1602,25 @@ declare namespace SendBird {
     hasNext: boolean;
     isLoading: boolean;
     includeEmpty: boolean;
-    order: 'latest_last_message' | 'chronological' | 'channel_name_alphabetical' | 'metadata_value_alphabetical';
+    order: "latest_last_message" | "chronological" | "channel_name_alphabetical" | "metadata_value_alphabetical";
     userIdsFilter: Array<string>; // DEPRECATED
     userIdsFilterExactMatch: boolean; // DEPRECATED
-    queryType: 'AND' | 'OR'; // DEPRECATED
+    queryType: "AND" | "OR"; // DEPRECATED
     userIdsExactFilter: Array<string>;
     userIdsIncludeFilter: Array<string>;
-    userIdsIncludeFilterQueryType: 'AND' | 'OR';
+    userIdsIncludeFilterQueryType: "AND" | "OR";
     nicknameContainsFilter: string;
     channelNameContainsFilter: string;
     customTypeFilter: string; // DEPRECATED
     customTypesFilter: Array<string>;
     customTypeStartsWithFilter: string;
     channelUrlsFilter: Array<string>;
-    superChannelFilter: 'all' | 'super' | 'nonsuper';
-    publicChannelFilter: 'all' | 'public' | 'private';
+    superChannelFilter: "all" | "super" | "nonsuper";
+    publicChannelFilter: "all" | "public" | "private";
     metadataOrderKeyFilter: string;
-    memberStateFilter: 'all' | 'joined_only' | 'invited_only' | 'invited_by_friend' | 'invited_by_non_friend';
-    hiddenChannelFilter: 'unhidden_only' | 'hidden_only' | 'hidden_allow_auto_unhide' | 'hidden_prevent_auto_unhide';
-    unreadChannelFilter: 'all' | 'unread_message';
+    memberStateFilter: "all" | "joined_only" | "invited_only" | "invited_by_friend" | "invited_by_non_friend";
+    hiddenChannelFilter: "unhidden_only" | "hidden_only" | "hidden_allow_auto_unhide" | "hidden_prevent_auto_unhide";
+    unreadChannelFilter: "all" | "unread_message";
     includeFrozen: boolean;
 
     setSearchFilter(fields: Array<string>, queryString: string): void;
@@ -1838,13 +1633,13 @@ declare namespace SendBird {
     hasNext: boolean;
     isLoading: boolean;
     includeEmpty: boolean;
-    order: 'chronological' | 'channel_name_alphabetical' | 'metadata_value_alphabetical';
+    order: "chronological" | "channel_name_alphabetical" | "metadata_value_alphabetical";
     channelNameContainsFilter: string;
     channelUrlsFilter: Array<string>;
     customTypesFilter: Array<string>;
     customTypeStartsWithFilter: string;
-    superChannelFilter: 'all' | 'super' | 'nonsuper';
-    membershipFilter: 'all' | 'joined';
+    superChannelFilter: "all" | "super" | "nonsuper";
+    membershipFilter: "all" | "joined";
     metadataOrderKeyFilter: string;
     includeFrozen: boolean;